+++ src/main/java/org/ala/biocache/web/ExploreController.java
@@ -57,17 +57,17 @@ public class ExploreController {
     private HashMap<String, List<Float>> addressCache = new HashMap<String, List<Float>>();
     private final String DEFAULT_LOCATION = "Clunies Ross St, Black Mountain, ACT";
     /** Mapping of radius in km to OpenLayers zoom level */
-    public final static HashMap<Integer, Integer> radiusToZoomLevelMap = new HashMap<Integer, Integer>();
+    public final static HashMap<Float, Integer> radiusToZoomLevelMap = new HashMap<Float, Integer>();
 	static {
-		radiusToZoomLevelMap.put(1, 14);
-        radiusToZoomLevelMap.put(5, 12);
-		radiusToZoomLevelMap.put(10, 11);
-		radiusToZoomLevelMap.put(50, 9);
+		radiusToZoomLevelMap.put(1f, 14);
+        radiusToZoomLevelMap.put(5f, 12);
+		radiusToZoomLevelMap.put(10f, 11);
+		radiusToZoomLevelMap.put(50f, 9);
 	}
 
     @RequestMapping(value = "/explore/your-area*", method = RequestMethod.GET)
 	public String yourAreaView(
-            @RequestParam(value="radius", required=false, defaultValue="5") Integer radius,
+            @RequestParam(value="radius", required=false, defaultValue="5") Float radius,
             @RequestParam(value="latitude", required=false, defaultValue="-35.27412f") Float latitude,
             @RequestParam(value="longitude", required=false, defaultValue="149.11288f") Float longitude,
             @RequestParam(value="address", required=false, defaultValue=DEFAULT_LOCATION) String address,
@@ -156,7 +156,7 @@ public class ExploreController {
      */
 	@RequestMapping(value = "/explore/download*", method = RequestMethod.GET)
 	public void yourAreaDownload(
-            @RequestParam(value="radius", required=false, defaultValue="10") Integer radius,
+            @RequestParam(value="radius", required=false, defaultValue="10") Float radius,
             @RequestParam(value="latitude", required=false, defaultValue="0f") Float latitude,
             @RequestParam(value="longitude", required=false, defaultValue="0f") Float longitude,
             @RequestParam(value="taxa", required=false, defaultValue="") String taxa, // comma separated list
@@ -206,7 +206,7 @@ public class ExploreController {
      */
     @RequestMapping(value = "/explore/species.json", method = RequestMethod.GET)
 	public void listSpeciesForHigherTaxa(
-            @RequestParam(value="radius", required=false, defaultValue="10") Integer radius,
+            @RequestParam(value="radius", required=false, defaultValue="10") Float radius,
             @RequestParam(value="latitude", required=false, defaultValue="0f") Float latitude,
             @RequestParam(value="longitude", required=false, defaultValue="0f") Float longitude,
             @RequestParam(value="taxa", required=false, defaultValue="") String taxa, // comma separated list
